Use original value types to get the common type.
When getting a common type for 2 values, use the actual values of the actual types
rather than the values which had been casted to longs. Also don't cast the
converted result to a long if we want it to actually be of the common type
as opposed to always being a long.

Fixes bug #2984.